+++ CHANGELOG.md
@@ -35,6 +35,9 @@ release.
 
 ## [Unreleased]
 
+### Fixed
+- Fixed default `limitQuality` option to `false` [#80](https://github.com/DOI-USGS/SpiceQL/pull/80)
+
 ## 1.2.0
 
 ### Added